+++ CHANGELOG.md
@@ -1,9 +1,32 @@
 # Changelog
 
+## [1.0.1] - 2025-02-24
+### Update
+- ga_core.py: add Annotations, Type Hints, Docstrings, comments,formatted 
+  code to meet PEP 8 rules.
+
+- ga_launch.py: add Anotations, Type Hints, Doctrings, Comments, formatted 
+  code to meet PEP 8 rules.
+
+- day-6-exercises_.py: naming conventions and whitespaces, documentation and
+  comments. code organization and modularity, error handling.
+
+### Added
+- taco_maker.py: taco maker app
+
+## [1.0.0] - 2025-02-11
+### Added
+- ga_core.py: add functions to add, edit, list, and remove items from
+  a grocery list
+  
+- ga_launch.py: add functions to interact with user to add, edit, and remove 
+  items from a grocery list
+
 ## [1.0.0] - 2025-02-11
 ### Added
 - ga_core.py: add functions to add, edit, list, and remove items from a grocery list
 - ga_launch.py: add functions to interact with user to add, edit, and remove items from a grocery list
+
 
 ### Removed
 - my_first_script.py